requirements.txt
flask==2.0.1
pyserial==3.5
werkzeug==2.0.3 